web/requirements.txt
Flask
psycopg-binary==3.2.1
psycopg==3.2.1
# psycopg2==3.2.1
SQLAlchemy==2.0.31
# Flask-SQLAlchemy==3.1.1
python-dotenv==1.1.0
pandas==2.2.2